python_model_deployment/requirements.txt
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
JayDeBeApi==1.2.3
joblib==1.0.1
PyJWT==2.1.0
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
pymongo==3.11.4
python-dotenv==0.17.1
scikit-learn==0.24.2
